+++ Elmish.XamarinForms.LiveUpdate/AssemblyInfo.fs
@@ -5,13 +5,13 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("Elmish.XamarinForms.LiveUpdate")>]
 [<assembly: AssemblyProductAttribute("Elmish.XamarinForms.LiveUpdate")>]
 [<assembly: AssemblyDescriptionAttribute("F# Functional App Dev Framework Live Update")>]
-[<assembly: AssemblyVersionAttribute("0.14.0")>]
-[<assembly: AssemblyFileVersionAttribute("0.14.0")>]
+[<assembly: AssemblyVersionAttribute("0.14.4")>]
+[<assembly: AssemblyFileVersionAttribute("0.14.4")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "Elmish.XamarinForms.LiveUpdate"
     let [<Literal>] AssemblyProduct = "Elmish.XamarinForms.LiveUpdate"
     let [<Literal>] AssemblyDescription = "F# Functional App Dev Framework Live Update"
-    let [<Literal>] AssemblyVersion = "0.14.0"
-    let [<Literal>] AssemblyFileVersion = "0.14.0"
+    let [<Literal>] AssemblyVersion = "0.14.4"
+    let [<Literal>] AssemblyFileVersion = "0.14.4"