+++ fscd.Tests/Tests.fs
@@ -331,7 +331,7 @@ if c 5 <> 5 then failwith "fail fail fail!"
 
     [<TestMethod>]
     member this.DateTime() =
-        SimpleTestCase "GenericThing" """
+        SimpleTestCase "DateTime" """
 let v1 = System.DateTime.Now
 let v2 = v1.Date
 let mutable v3 = System.DateTime.Now
@@ -340,13 +340,26 @@ let v4 = v3.Date
 
     [<TestMethod>]
     member this.LocalMutation() =
-        SimpleTestCase "GenericThing" """
+        SimpleTestCase "LocalMutation" """
 let f () = 
     let mutable x = 1
     x <- x + 1
     x <- x + 1
     x
 if f() <> 3 then failwith "fail fail!" 
+        """
+
+
+    [<TestMethod>]
+    member this.LetRecSmoke() =
+        SimpleTestCase "LetRecSmoke" """
+let even a = 
+    let rec even x = (if x = 0 then true else odd (x-1))
+    and odd x = (if x = 0 then false else even (x-1))
+    even a
+
+if even 11 then failwith "fail!" 
+if not (even 10) then failwith "fail fail!" 
         """